Split EmojiDataSource
