+++ vector/src/main/java/im/vector/riotx/features/reactions/EmojiRecyclerAdapter.kt
@@ -30,6 +30,7 @@ import androidx.recyclerview.widget.RecyclerView
 import androidx.transition.AutoTransition
 import androidx.transition.TransitionManager
 import im.vector.riotx.R
+import im.vector.riotx.features.reactions.data.EmojiDataSource
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.GlobalScope
 import kotlinx.coroutines.launch
