+++ vector/src/main/java/im/vector/riotx/features/reactions/data/EmojiDataSource.kt
@@ -13,11 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package im.vector.riotx.features.reactions
+package im.vector.riotx.features.reactions.data
 
 import android.content.Context
-import com.squareup.moshi.Json
-import com.squareup.moshi.JsonClass
 import com.squareup.moshi.Moshi
 import im.vector.riotx.R
 
@@ -31,32 +29,6 @@ class EmojiDataSource(val context: Context) {
             val jsonAdapter = moshi.adapter(EmojiData::class.java)
             val inputAsString = input.bufferedReader().use { it.readText() }
             this.rawData = jsonAdapter.fromJson(inputAsString)
-           // this.rawData = mb.fr(InputStreamReader(it), EmojiData::class.java)
-        }
-    }
-    @JsonClass(generateAdapter = true)
-    data class EmojiData(val categories: List<EmojiCategory>,
-                         val emojis: Map<String, EmojiItem>,
-                         val aliases: Map<String, String>)
-
-    @JsonClass(generateAdapter = true)
-    data class EmojiCategory(val id: String, val name: String, val emojis: List<String>)
-
-    @JsonClass(generateAdapter = true)
-    data class EmojiItem(
-            @Json(name = "a") val name: String,
-            @Json(name = "b") val unicode: String,
-            @Json(name = "j") val keywords: List<String>?,
-            val k: List<String>?) {
-
-        var _emojiText: String? = null
-
-        fun emojiString() : String {
-            if (_emojiText == null) {
-                val utf8Text = unicode.split("-").joinToString("") { "\\u$it" } // "\u0048\u0065\u006C\u006C\u006F World"
-               _emojiText = fromUnicode(utf8Text)
-            }
-            return _emojiText!!
         }
     }
 
@@ -72,20 +44,4 @@ class EmojiDataSource(val context: Context) {
             return text.toString()
         }
     }
-
-//    name: 'a',
-//    unified: 'b',
-//    non_qualified: 'c',
-//    has_img_apple: 'd',
-//    has_img_google: 'e',
-//    has_img_twitter: 'f',
-//    has_img_emojione: 'g',
-//    has_img_facebook: 'h',
-//    has_img_messenger: 'i',
-//    keywords: 'j',
-//    sheet: 'k',
-//    emoticons: 'l',
-//    text: 'm',
-//    short_names: 'n',
-//    added_in: 'o',
 }
